CloudFunctions/media_agent_function/requirements.txt
python-dotenv
requests
google-cloud-firestore
vertexai
functions-framework
google-cloud
